feat multi network: init network select button
